// backend/src/types/shared.ts
export enum WorkStatus {
  DRAFT = 'DRAFT',
  PUBLISHED = 'PUBLISHED',
}

export enum WorkCategory {
  BRANDING = 'BRANDING',
  WEBSITE = 'WEBSITE',
  SYSTEM = 'SYSTEM',
  ADVERTISING = 'ADVERTISING',
  THREED = 'THREED',
  OTHER = 'OTHER',
  DESIGN = 'DESIGN',
  SAAS = 'SAAS',
  PHOTOGRAPHY = 'PHOTOGRAPHY',
}

export enum WorkType {
  DESIGN = 'DESIGN',
  DEVELOPMENT = 'DEVELOPMENT',
  MARKETING = 'MARKETING',
  ANIMATION = 'ANIMATION',
  PHOTOGRAPHY = 'PHOTOGRAPHY',
  OTHER = 'OTHER',
}